fix: reformat login structure
