+++ src/App.js
@@ -1,92 +1,43 @@
-import React from 'react'
-import { Experiment, jsPsych } from 'jspsych-react'
-import { Redirect, BrowserRouter as Router, Switch, Route } from 'react-router-dom';
-import { tl } from './timelines/main'
-import { MTURK, FIREBASE, IS_ELECTRON } from './config/main'
-import './App.css'
-import 'bootstrap/dist/css/bootstrap.css'
-import '@fortawesome/fontawesome-free/css/all.css'
-import { getTurkUniqueId, sleep } from './lib/utils'
-import { rt_categorize_html } from './lib/rt-categorize-html'
-import { addToFirebase } from "./firebase.js";
-import Login from './login'
+import React, { useState, useEffect } from "react";
+import "./App.css";
+import "bootstrap/dist/css/bootstrap.css";
+import "@fortawesome/fontawesome-free/css/all.css";
+import Login from "./components/Login";
+import JsPsychExperiment from "./components/JsPsychExperiment";
+import { MTURK, IS_ELECTRON } from "./config/main";
+import { getTurkUniqueId } from "./lib/utils";
+import LoginContext from "./contexts/LoginContext";
 
+function App() {
+  // Variables for login context
+  const [loggedIn, setLogin] = useState(false);
+  const login = () => setLogin(true);
 
+  const [ipcRenderer, setRenderer] = useState(false);
+  const [psiturk, setPsiturk] = useState(false);
 
-var LOGGEDIN = false
-const isElectron = IS_ELECTRON//!MTURK
-let ipcRenderer = false;
-let psiturk = false
-if (isElectron) {
-  const electron = window.require('electron');
-  ipcRenderer  = electron.ipcRenderer;
-} else if (MTURK){
-  /* eslint-disable */
-  window.lodash = _.noConflict()
-  psiturk = new PsiTurk(getTurkUniqueId(), '/complete')
-  /* eslint-enable */
+  // All Mturk and other login logic will go here
+  useEffect(() => {
+    if (IS_ELECTRON) {
+      const electron = window.require("electron");
+      setRenderer(electron.ipcRenderer);
+    } else if (MTURK) {
+      /* eslint-disable */
+      window.lodash = _.noConflict();
+      setPsiturk(new PsiTurk(getTurkUniqueId(), "/complete"));
+      /* eslint-enable */
+    }
+  }, []);
+
+  return (
+    <LoginContext.Provider value={{ loggedIn, login }}>
+      {loggedIn ? (
+        <JsPsychExperiment ipcRenderer={ipcRenderer} psiturk={psiturk} />
+      ) : (
+        <Login />
+      )}
+    </LoginContext.Provider>
+  );
 }
 
-
-
-class ExpStart extends React.Component {
-  render() {
-    console.log("Outside Turk:", jsPsych.turk.turkInfo().outsideTurk)
-    console.log("Turk:", MTURK)
-    jsPsych.plugins['rt-categorize-html'] = rt_categorize_html();
-    
-    return (
-      <div className="App">
-        <Experiment settings={{
-          timeline: tl,
-          on_data_update: (data) => {
-            //firebase 
-            if(FIREBASE){
-              addToFirebase(data);
-            }
-            else if ( ipcRenderer ) {
-              ipcRenderer.send('data', data)
-            }
-            else if (psiturk) {
-                psiturk.recordTrialData(data)
-            }
-          },
-          on_finish: (data) => {
-            
-            if ( ipcRenderer ) {
-              ipcRenderer.send('end', 'true')
-            }
-            else if (psiturk) {
-              const completePsiturk = async () => {
-                psiturk.saveData()
-                await sleep(5000)
-                psiturk.completeHIT()
-              }
-              completePsiturk()
-            }
-          },
-        }}
-        />
-      </div>
-    );
-  }
-}
-
-class App extends React.Component {
-  render() {
-    return (
-      <>
-        <Router>
-          <Switch>
-          <Route exact path="/">
-            {LOGGEDIN ? <Redirect to="/experiment" /> : <Login/>}
-          </Route>
-            <Route path='/experiment' component={ExpStart} />
-          </Switch>
-        </Router>
-      </>
-    );
-  }
-}
-
-export default App
+export default App;
