+++ src/contexts/LoginContext.js
@@ -1,0 +1,5 @@
+import { createContext } from 'react'
+
+const LoginContext = createContext(false);
+
+export default LoginContext
